Fix #441 by reverting 9569e61. Fix memory leak in session manager instead.
When notebookcellmodelmanager.reset() is called with no arguments,
it should keep the cells as is and just rebuild the cell map,
which was broken by 9569e61.

After this fix, if one want to reset the cells to empty, [] needs
to be passed explicitly. As a result, when session model manager
resets the notebook model, it now calls reset([]) instead of reset().