Add 'beaker-sandbox' class to DOM elements that need Beaker styles.
This change should have no effect on desktop beaker, but will facilitate
smoothly integrating beaker with containing applications.

When we embed Beaker in external systems, we want to avoid the possibility
to interfere with styles outside beaker's own rendering area, and giving
a class name to elements which contain Beaker content allows
us to differentiate in the unified stylesheets.

The Beaker dialogs (being Bootstrap dialogs)
are attached directly to the body element, and need this class.

In the desktop version, EVERY element gets beaker styles. Although
index_template.html will probably not be used for the "embedded version" of
beaker (when it becomes available), for consistency let's also mark the
root ng-view element of desktop beaker as containing beaker content by
using the same css class name.