Take off aggressive scroll correcting
The old code not only corrected the scroll after $timeout,
but it added a listener to on scroll scroll back to 0.
This wasn't really a problem, it would just cancel out the user's
first scroll (minute), but in order to add the ability to scroll
a cell into view, we'll need to make this function slightly less
aggressive.