style: Fix markdown output overlap with menu.
The issue, was the fact that we have horizontal menu of cell actions
that are absolutely positioned on top of the output cell. Due to the
positioning the menu, the markdown cell output did not flow around the menu
element, thus causing a visual overlap. The CSS2 spec did introduce the
concept of a `::first-line` pseudo selector, however the rules that can be
applied are limited to a small subset of css rules [1]. None of which
can limit the width of the first line. The solution that I opted for
instead was to take advantage of a CSS3 pseudo selector `:before` to
insert a "fake" element that takes up the same size as the menu that is
out of the flow. into the output element to cause the output text
to reflow around

[1] https://developer.mozilla.org/en-US/docs/Web/CSS/::first-line

Fixes #1532