Allow users to enter a escape $
Before this fix \$ was shown as \$ in the markdown output.
As part of the parsing step for marked we are now replacing instances of
`\$` with `$`. This step runs after the katex interpolation so there is
no need to be concerned with the `$` being interpolated into a new LateX
section.

Fixes #1428