Lazily instantiate codemirror instances
By taking advantage of the fact that a user can not interact
with codemirror instance until they are in the viewport, we can be
smart and only set the codemirror editors up once a user scrolls them into view.

For small notebooks this change has a limited impact, however for large
notebooks with 100s of code cells this change results in a significant
performance boost to the application, especially in the initial time to
render metric for notebooks.

ref #1546
Fixes #178