Fix padding application on language picker.
Prior to this change, we were using the language picker to apply padding
to the code editor. This choice lead to a css bug with the right menu in
advanced mode (the height being too large).

The fix in this case was to change the padding rule to be on the editor
node as compared to a node that happens to be near to it.

Made visible in 5d8677459665703a59992cb199164d2f174561ea

Code initally added in dba4979368193a37886420117ca77e45d17f114f

Fixes #1682