Fix delayed spinner hide in advanced mode.
Angular prevents hiding an element while an animation is happening on
the element. In this case we were trying to hide an element that was
"spinning" and since it had a 2 second animation duration, the ng-if would only
act on the element after the animation cycle had completed.

By putting a containing div on the ng-if directive we work around this
limitation. In someways this is a hack.

Fixes #1771