Fix z-index cell stacking bug.
Fixes #1815

the z-index: 10 and node stacking lead to this bug, initially added in
https://github.com/twosigma/beaker-notebook/commit/eb17a7013701954ee03ad9544e82c147cfd8b647

But was a new "fix" from an earlier "fix" in #951

The origin of the issue, is simply in advanced mode, the editor is on
top (in the DOM) of the menu, and it needs to be z stacked below it.

This solution is only focued on the relationship between the menu and
the code-cell and nothing else.