Fix z-index cell stacking bug.
Fixes #1815

the z-index: 10 and node stacking lead to this bug, initially added in
https://github.com/twosigma/beaker-notebook/commit/eb17a7013701954ee03ad9544e82c147cfd8b647

But was a new "fix" from an earlier "fix" in #951

The origin of the issue, is simply in advanced mode, the combination of z-index, position attributes, DOM order, and codemirror z-index was messing up the stacking order of all of the elements. The solution was to reorder the DOM nodes and to remove the z-index of the codemirror pre.

ref:
https://developer.mozilla.org/en-US/docs/Web/Guide/CSS/Understanding_z_index/Stacking_context_example_1