Remove zip from repo
This change should prevent the entire class of issues related to #1746.
From this point forward when someone adds, removes, or updates a
dependency in the repo it will show up in the diff, making it easy to
spot when the incorrect files get added or deleted.

The initial motivation for the zip file was because there was so
many files from mathjax, but after moving to katex and updating our
gitignore to not track unused vendor files, the total number of files that have been added to the repo is
644.