Close dropdowns when a new dropdown is opened.
By listening to the custom event that is emitted by bootstrap when a
dropdown is opened we can then close all of the open bootstrap dropdown
menus.

The reason we have to manually listen for this is because since we are
promoting the dropdown outside of the dropdown toggle button parent the
bootstrap plugin is unable to find the menu to close.

Fixes #1842