fix plots need more lines #2029
I made predefined time an values more detailed and decreased desired step for time axe to 105 from 150.

scottdraves commented 19 hours ago
Thanks.
This is a step in the right direction. But still, the lines are never closer than about 104 pixels.
can you make them go down to 50?

This is not a good idea. Time labels may overlap in some cases.