fix codemirror should unhighlight on blur #2261
I found that highlighting not working if selection in an editor is not changed. I has added event listeners on "blur" and "focus" for a solving this problem.
On "blur" selection is changed and  "cursorActivity" event fired. On "focus" only "cursorActivity" event fired.

Notebook for tests - https://www.dropbox.com/s/lys4nixenlzaci9/2261.bkr?dl=0