scroll on code cell create #3039
My proposal is to using standard codemirror focus mechanism instead hand-made function scrollIn. When we are creating codecell (or markdown cell) we are calling codemirror focus function. This function moves offscreen codemiiror in the center of window. This is not full expected behaviour for this issue but I have opinion that using standard tested mechanism is a better solution. What is your opinion?