timeseries labels and lines should be at beginning of interval #3061
I adopted algorithm for for days, months, and years. I choose the beginning of the next range (day, month or year), if the calculated value closer to the  end of the period, otherwise - end of the previous range. This is also fix for the https://github.com/twosigma/beaker-notebook/issues/3052