plot hover circle inconsistent #3025
i don't see what you mean about the problem with using the clicked location.
the image above looks fine. what's the problem?

DONE

also still missing: the item you are hovered over should highlight (get fatter).
similar to the legend, if you move the mouse quickly over the item it should not respond.
this is true of the tooltip in general.

Problem was only for stems - FIXED. Bars and Points uses svgfilter