support 64 bit integers in tables and autotranslation #3332
scottdraves commented an hour ago
OK this is a good start but the issue is much larger. Eventually we are going to need the object format specified in the issue. But we can get there in some smaller steps.

First step, it should work with Longs in addition to BigIntegers. DONE

Also, add your tutorial as a separate notebook to the tutorials. DONE