vim edit mode broken #3475
ESC in code cell does not exit insert mode and return to movement mode. FIXED

also, the solid cursor blocks the visibility of the character that it covers.
it should be inverse.

I have used css property 'opacity' for a fix this problem.

Also I removed 'width: 0.5em;' for the 'CodeMirror.cm-fat-cursor div.CodeMirror-cursor'  In this case we have best cursor visualisation for code cells with monospaced font. But cursor for the markdown cell is tiny.