Updated run_gradle function in two ways.
First, it is now called `install_gradle` which matches the convention used in setupbase_jp.

Second, the `run` function from setupbase_jp now has a line that calls `which` to see if the command you want to run is in your path. If the command is not found, it will fail. `gradlew` is packaged with beakerx so it will not be in a user's path. To address, the full path to gradlew is used instead as `which` will make sure the full path exists and then return the same full path, letting the command work properly.