improve STIL example (#6722)
Fix the STIL example so that it accesses the StarTable object
cell values directly rather than just reading all the cells
into an internal array and accessing that.

The StarTable object requires random access for this to work;
tables read from CSV files are not random-access by default,
so the Tables.randomTable method has to get called on it.

In fact this is just moving the copy-to-internal-array step
out of the groovy and into the STIL library, so in this case
it doesn't make a huge amount of difference.  However, if the
input file was e.g. a FITS file, it would mean that data access
was strictly on-demand to the mapped disk file, and so potentially
much more efficient, especially for large files.